+++ lib/czmq-ffi-gen/versions.rb
@@ -1,6 +1,3 @@
-require_relative "gem_version"
-require_relative "libzmq"
-
 module CZMQ
   module FFI
     # CZMQ library version
@@ -10,10 +7,6 @@ module CZMQ
     LIBRARY_VERSION = CZMQ_VERSION
 
     # ZMQ library version
-    ZMQ_VERSION = begin
-      version = Array.new(3) { ::FFI::MemoryPointer.new(:int) }
-      LibZMQ.zmq_version(*version)
-      version.map { |n| n.get_int(0) }.join(".")
-    end
+    ZMQ_VERSION = LibZMQ::VERSION
   end
 end