+++ spec/czmq-ffi-gen/czmq/ffi/czmq_ffi_extensions_spec.rb
@@ -1,0 +1,9 @@
+require_relative "spec_helper"
+
+describe CZMQ::FFI do
+  describe ".has_draft?" do
+    it 'returns boolean' do
+      assert_includes [true, false], CZMQ::FFI.has_draft?
+    end
+  end
+end