+++ lib/czmq-ffi-gen/gem_version.rb
@@ -1,5 +1,5 @@
 module CZMQ
   module FFI
-    GEM_VERSION = "0.16.0"
+    GEM_VERSION = "0.16.1"
   end
 end