+++ CHANGES.md
@@ -4,6 +4,8 @@
 * update CZMQ low-level binding
 * remove support for CZMQ 3
 * add CZMQ::FFI::LibZMQ.zmq_has, .has?(cap), .has_curve?, .has_draft?
+* add CZMQ::FFI:has_draft?
+* add CZMQ::FFI::LibZMQ::VERSION
 
 0.15.0 (03/04/2018)
 -----